Break logic on charities into separate file
